cbuild: remove system_users/system_groups metadata
Instead, anything that needs a user/group created for its build
can do so by creating a sysusers.conf file and putting it in the
template directory. It will be picked up early in the build and
its users/groups will be created.

+++ src/cbuild/hooks/prepare/001_prepare_users.py
@@ -2,7 +2,7 @@
 
 from cbuild.core import chroot, paths
 
-import shlex
+import shutil
 
 
 def invoke(pkg):
@@ -10,74 +10,16 @@ def invoke(pkg):
     if pkg.stage < 1:
         return
 
+    if not (pkg.template_path / "sysusers.conf").exists():
+        return
+
     # assert this, should always be true...
     if not (paths.bldroot() / "usr/bin/sd-sysusers").exists():
         pkg.error("sd-sysusers not present in chroot")
 
-    glist = []
-    ulist = []
-
-    # add new groups
-    for g in pkg.system_groups:
-        gl = g.split(":")
-        if len(gl) == 1:
-            gname = g
-            gid = "-"
-        else:
-            gname = gl[0]
-            gid = gl[1]
-        glist.append(f"g {gname} {gid}")
-
-    # add new users
-    for u in pkg.system_users:
-        if isinstance(u, dict):
-            uname = u["name"]
-            uid = u["id"]
-            if not uid:
-                uid = "-"
-            # home dir
-            if "home" in u:
-                uhm = u["home"]
-            else:
-                uhm = "/tmp"
-            # comment
-            if "desc" in u:
-                ucmt = u["desc"]
-            else:
-                ucmt = f"{uname} user"
-            # shell
-            if "shell" in u:
-                ush = u["shell"]
-            else:
-                ush = "/bin/sh"
-            # primary gid
-            if "pgroup" in u:
-                gn = u["pgroup"]
-            else:
-                gn = None
-        else:
-            ul = u.split(":")
-            if len(ul) == 1:
-                uname = u
-                uid = "-"
-            else:
-                uname = ul[0]
-                uid = ul[1]
-            # remaining info is automatic
-            gn = None
-            uhm = "/tmp"
-            ucmt = f"{uname} user"
-            ush = "/bin/sh"
-        if gn:
-            uid = f"{uid}:{gn}"
-        ulist.append(f"u {uname} {uid} {shlex.quote(ucmt)} {uhm} {ush}")
-
-    # generate sysusers file
-    with open(paths.bldroot() / "usr/lib/sysusers.d/cbuild.conf", "w") as outf:
-        for gl in glist:
-            outf.write(f"{gl}\n")
-        for ul in ulist:
-            outf.write(f"{ul}\n")
+    bp = paths.bldroot() / "usr/lib/sysusers.d"
+    bp.mkdir(exist_ok=True)
+    shutil.copyfile(pkg.template_path / "sysusers.conf", bp / "cbuild.conf")
 
     # delete potential shadow so sysusers does not fail
     (paths.bldroot() / "etc/shadow").unlink(missing_ok=True)
